Remove unneeded assertify_proc_macros crate
The actual definitions of the `testify` and `assertify` macros were in a
sub-crate called `assertify_proc_macros`. There was no need to do that
as long as the tests were defined in the tests/ directory rather than in
src/lib.rs, so this removes the sub-crate.
