--- assertify_proc_macros/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "assertify_proc_macros"
-version = "0.6.1"
-authors = ["Daniel Parks <oss-assertify@demonhorse.org>"]
-description = "Actual macro code for https://github.com/danielparks/assertify. Do not use this directly."
-homepage = "https://github.com/danielparks/assertify"
-repository = "https://github.com/danielparks/assertify"
-license = "MIT OR Apache-2.0"
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = { version = "1.0", features = ["full", "extra-traits"] }
