+++ tests/trybuild.rs
@@ -1,0 +1,5 @@
+#[test]
+fn trybuild_tests() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/trybuild-failures/*.rs");
+}
